requirements.txt
# TaxAI Backend Requirements

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database & ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# AI & LLM
openai==1.3.5
crewai==0.11.0
crewai-tools==0.2.0
langchain==0.0.339
langchain-openai==0.0.2

# HTTP Requests
requests==2.31.0

# Data Processing
pandas==2.1.3
pydantic==2.5.0
pydantic-settings==2.1.0

# Environment
python-dotenv==1.0.0

# CORS
python-cors==1.0.0
